format code, optimize code, update doc,

+++ src/main/java/com/d_project/qrcode/BitBuffer.java
@@ -2,7 +2,8 @@ package com.d_project.qrcode;
 
 /**
  * BitBuffer
- * @author Kazuhiko Arase 
+ *
+ * @author Kazuhiko Arase
  */
 class BitBuffer {
 
@@ -27,31 +28,30 @@ class BitBuffer {
   public String toString() {
     StringBuilder buffer = new StringBuilder();
     for (int i = 0; i < getLengthInBits(); i++) {
-      buffer.append(get(i)? '1' : '0');
+      buffer.append(get(i) ? '1' : '0');
     }
     return buffer.toString();
   }
 
   private boolean get(int index) {
-    return ( (buffer[index / 8] >>> (7 - index % 8) ) & 1) == 1;
+    return ((buffer[index / 8] >>> (7 - index % 8)) & 1) == 1;
   }
 
   public void put(int num, int length) {
     for (int i = 0; i < length; i++) {
-      put( ( (num >>> (length - i - 1) ) & 1) == 1);
+      put(((num >>> (length - i - 1)) & 1) == 1);
     }
   }
 
   public void put(boolean bit) {
-
     if (length == buffer.length * 8) {
       byte[] newBuffer = new byte[buffer.length + inclements];
       System.arraycopy(buffer, 0, newBuffer, 0, buffer.length);
-      buffer = newBuffer;      
+      buffer = newBuffer;
     }
 
     if (bit) {
-      buffer[length / 8] |= (0x80 >>> (length % 8) );
+      buffer[length / 8] |= (byte) (0x80 >>> (length % 8));
     }
 
     length++;
